refactor(notice): 게시글 내 중요 공지 Box 컴포넌트로 분리 (#66)
