+++ src/icons/Alarm.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+import alarmIcon from '@assets/svgs/notice/alarm.svg?react';
+
+export const AlarmIcon = styled(alarmIcon)`
+  width: 1.5rem;
+  height: fit-content;
+`;
